+++ simter-file-rest-webflux/src/main/kotlin/tech/simter/file/rest/webflux/handler/UploadHandler.kt
@@ -2,8 +2,8 @@ package tech.simter.file.rest.webflux.handler
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus.CREATED
-import org.springframework.http.HttpStatus.FORBIDDEN
-import org.springframework.http.MediaType.*
+import org.springframework.http.MediaType.MULTIPART_FORM_DATA
+import org.springframework.http.MediaType.MULTIPART_MIXED
 import org.springframework.http.codec.multipart.FilePart
 import org.springframework.http.codec.multipart.Part
 import org.springframework.stereotype.Component
@@ -20,6 +20,7 @@ import tech.simter.exception.PermissionDeniedException
 import tech.simter.file.core.FileDescriber
 import tech.simter.file.core.FileService
 import tech.simter.file.core.FileUploadSource
+import tech.simter.reactive.web.Utils
 
 /**
  * The [HandlerFunction] for upload file.
@@ -89,10 +90,9 @@ class UploadHandler @Autowired constructor(
           .flatMap { status(CREATED).bodyValue(it) }
       }
     }
-    return uploadResult.onErrorResume(PermissionDeniedException::class.java) {
-      if (it.message.isNullOrEmpty()) status(FORBIDDEN).build()
-      else status(FORBIDDEN).contentType(TEXT_PLAIN).bodyValue(it.message!!)
-    }
+    return uploadResult
+      // permission denied
+      .onErrorResume(PermissionDeniedException::class.java, Utils::responseForbiddenStatus)
   }
 
   companion object {