+++ simter-file-service-impl/src/main/kotlin/tech/simter/file/impl/FileServiceImpl.kt
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service
 import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 import tech.simter.exception.NotFoundException
-import tech.simter.file.BASE_DATA_DIR
+import tech.simter.file.*
 import tech.simter.file.core.*
 import tech.simter.kotlin.data.Page
 import tech.simter.reactive.security.ReactiveSecurityService
@@ -29,6 +29,7 @@ import java.util.*
  */
 @Service
 class FileServiceImpl @Autowired constructor(
+  private val fileAuthorizer: FileAuthorizer,
   @Value("\${$BASE_DATA_DIR}")
   private val baseDir: String,
   @Value("\${simter-file.verify-real-file-size: true}")
@@ -42,7 +43,7 @@ class FileServiceImpl @Autowired constructor(
   val filePathGenerator: FilePathGenerator,
   val securityService: ReactiveSecurityService
 ) : FileService {
-  private final val logger = LoggerFactory.getLogger(FileServiceImpl::class.java)
+  private val logger = LoggerFactory.getLogger(FileServiceImpl::class.java)
   private val basePath: Path = Paths.get(baseDir)
 
   override fun findPage(
@@ -51,12 +52,15 @@ class FileServiceImpl @Autowired constructor(
     limit: Optional<Int>,
     offset: Optional<Long>
   ): Mono<Page<FileStore>> {
-    return fileDao.findPage(
-      moduleMatcher = moduleMatcher,
-      search = search,
-      limit = limit.orElse(defaultFindPageLimit),
-      offset = offset.orElse(0)
-    )
+    return fileAuthorizer.verifyHasPermission(moduleMatcher.module, OPERATION_READ)
+      .then(Mono.defer {
+        fileDao.findPage(
+          moduleMatcher = moduleMatcher,
+          search = search,
+          limit = limit.orElse(defaultFindPageLimit),
+          offset = offset.orElse(0)
+        )
+      })
   }
 
   override fun findList(
@@ -64,79 +68,102 @@ class FileServiceImpl @Autowired constructor(
     search: Optional<String>,
     limit: Optional<Int>
   ): Flux<FileStore> {
-    return fileDao.findList(
-      moduleMatcher = moduleMatcher,
-      search = search,
-      limit = if (limit.isPresent) limit else Optional.of(defaultFindListLimit)
-    )
+    return fileAuthorizer.verifyHasPermission(moduleMatcher.module, OPERATION_READ)
+      .thenMany(Flux.defer {
+        fileDao.findList(
+          moduleMatcher = moduleMatcher,
+          search = search,
+          limit = if (limit.isPresent) limit else Optional.of(defaultFindListLimit)
+        )
+      })
   }
 
-  /** Get the context user name */
-  private fun getCurrentUser(): Mono<String> {
+  /** Get the context username */
+  internal fun getCurrentUser(): Mono<String> {
     return securityService.getAuthenticatedUser()
       .map { if (it.isPresent) it.get().name else "System" }
   }
 
+  /** Get the current timestamp */
+  internal fun getCurrentTimestamp(): OffsetDateTime {
+    return OffsetDateTime.now()
+  }
+
+  /** create a random uuid */
+  internal fun generateUUID(): UUID {
+    return UUID.randomUUID()
+  }
+
   override fun upload(describer: FileDescriber, source: FileUploadSource): Mono<String> {
-    // unique markup
-    val ts = Optional.of(OffsetDateTime.now())
-    val uuid = Optional.of(UUID.randomUUID())
+    // generate unique markup
+    val ts = Optional.of(getCurrentTimestamp())
+    val uuid = Optional.of(generateUUID())
 
-    // 1. generate file path
-    val filePath = filePathGenerator.resolve(describer = describer, ts = ts, uuid = uuid)
-    val targetFile = basePath.resolve(filePath)
-    val parentDir = targetFile.toFile().parentFile
-    if (!parentDir.exists()) {
-      parentDir.mkdirs()
-      logger.info("create directory '{}'", parentDir)
-    }
+    return fileAuthorizer
+      // 1. verify permission
+      .verifyHasPermission(describer.module, OPERATION_CREATE)
+      .then(Mono.defer {
+        // 2. save physical file
+        // 2.1. generate file path
+        val filePath = filePathGenerator.resolve(describer = describer, ts = ts, uuid = uuid)
+        val targetFile = basePath.resolve(filePath)
+        val parentDir = targetFile.toFile().parentFile
+        if (!parentDir.exists()) {
+          parentDir.mkdirs()
+          logger.info("create directory '{}'", parentDir)
+        }
 
-    // 2. store file data to physical disk
-    val transferTo: Mono<Void> = when (source) {
-      is FileUploadSource.FromFilePart -> source.value.transferTo(targetFile).doOnSuccess {
-        // verify real file size
-        if (verifyRealFileSize && targetFile.toFile().length() != describer.size)
-          throw IllegalArgumentException("specific upload file size not match the real file size " + targetFile.toFile().length() + "|" + describer.size)
-      }
-      is FileUploadSource.FromResource -> {
-        // zero-copy for disk file
-        FileChannel.open(targetFile, CREATE_NEW, WRITE)
-          .transferFrom(source.value.readableChannel(), 0, source.value.contentLength())
-        Mono.empty()
-      }
-      is FileUploadSource.FromDataBufferPublisher -> {
-        val channel = AsynchronousFileChannel.open(targetFile, CREATE_NEW, WRITE)
-        DataBufferUtils.write(source.value, channel)
-          .map {
-            DataBufferUtils.release(it)
-            it
+        // 2.2. store file data to physical disk
+        val transferTo: Mono<Void> = when (source) {
+          is FileUploadSource.FromFilePart -> source.value.transferTo(targetFile).doOnSuccess {
+            // verify real file size
+            if (verifyRealFileSize && targetFile.toFile().length() != describer.size)
+              throw IllegalArgumentException("specific upload file size not match the real file size " + targetFile.toFile().length() + "|" + describer.size)
           }
-          .doOnTerminate { channel.close() }
-          .then()
-      }
-    }.doOnSuccess { logger.info("transfer file data to target file '{}'", targetFile) }
-
-    return transferTo.then(Mono.defer {
-      getCurrentUser().flatMap { creator ->
-        // 3. generate file id
-        fileIdGenerator.nextId(ts = ts, uuid = uuid)
-          .flatMap {
-            // 4. store file information to database and return the id
-            fileDao.create(FileStore.Impl(
-              id = it,
-              module = describer.module,
-              name = describer.name,
-              type = describer.type,
-              size = describer.size,
-              path = filePath.toString(),
-              creator = creator,
-              createOn = ts.get(),
-              modifier = creator,
-              modifyOn = ts.get()
-            ))
+          is FileUploadSource.FromResource -> {
+            // zero-copy for disk file
+            FileChannel.open(targetFile, CREATE_NEW, WRITE)
+              .transferFrom(source.value.readableChannel(), 0, source.value.contentLength())
+            Mono.empty()
           }
+          is FileUploadSource.FromDataBufferPublisher -> {
+            val channel = AsynchronousFileChannel.open(targetFile, CREATE_NEW, WRITE)
+            DataBufferUtils.write(source.value, channel)
+              .map {
+                DataBufferUtils.release(it)
+                it
+              }
+              .doOnTerminate { channel.close() }
+              .then()
+          }
+        }
+
+        // return the save-to-path
+        transferTo.doOnSuccess { logger.info("transfer file data to target file '{}'", targetFile) }
+          .thenReturn(filePath)
+      })
+      .flatMap { filePath ->
+        // get current user
+        getCurrentUser().flatMap { creator ->
+          // 3. generate file id
+          fileIdGenerator.nextId(ts = ts, uuid = uuid)
+            .flatMap { id ->
+              // 4. save to database and return the id
+              fileDao.create(FileStore.Impl(
+                id = id,
+                module = describer.module,
+                name = describer.name,
+                type = describer.type,
+                size = describer.size,
+                path = filePath.toString(),
+                creator = creator,
+                createOn = ts.get(),
+                modifier = creator,
+                modifyOn = ts.get()
+              ))
+            }
+        }
       }
-    })
   }
 
   override fun update(
@@ -145,11 +172,12 @@ class FileServiceImpl @Autowired constructor(
     source: Optional<FileUploadSource>
   ): Mono<Void> {
     // unique markup
-    val ts = Optional.of(OffsetDateTime.now())
-    val uuid = Optional.of(UUID.randomUUID())
+    val ts = Optional.of(getCurrentTimestamp())
+    val uuid = Optional.of(generateUUID())
 
     return fileDao.get(id)
       .switchIfEmpty(Mono.error(NotFoundException("no file to update was found!")))
+      .delayUntil { fileAuthorizer.verifyHasPermission(it.module, OPERATION_UPDATE) } // verify permission
       .flatMap { fileStore ->
         var size = fileStore.size
         val fileDescriber = FileDescriber.Impl(
@@ -217,33 +245,61 @@ class FileServiceImpl @Autowired constructor(
   }
 
   override fun download(id: String): Mono<FileDownload> {
-    return fileDao.get(id)
+    return fileDao
+      // get FileStore
+      .get(id)
+      // verify permission
+      .delayUntil { fileAuthorizer.verifyHasPermission(it.module, OPERATION_READ) }
+      // convert to FileDownload
       .map { FileDownload.from(it, basePath) }
   }
 
   override fun delete(vararg ids: String): Mono<Int> {
-    return fileDao.findById(*ids).collectList().flatMap { list ->
-      fileDao.delete(*ids).flatMap {
-        for (fileStore in list) {
+    return fileDao.findById(*ids)
+      .collectList()
+      // convert empty list to empty mono
+      .flatMap { if (it.isEmpty()) Mono.empty<List<FileStore>>() else Mono.just(it) }
+      // verify permission
+      .delayUntil { list ->
+        if (list.isEmpty()) Mono.empty<List<FileStore>>()
+        else Mono.`when`(list.map { fileAuthorizer.verifyHasPermission(it.module, OPERATION_DELETE) })
+      }
+      // delete db records
+      .flatMap { list -> fileDao.delete(*ids).map { Pair(list, it) } }
+      // delete physical files
+      .doOnNext {
+        for (fileStore in it.first) {
           val file = Paths.get(baseDir, fileStore.path).toFile()
           if (file.delete()) logger.info("delete file '{}'", file.absolutePath)
         }
-
-        Mono.just(it)
       }
-    }
+      .map { it.second }
+      .switchIfEmpty(Mono.just(0))
   }
 
   override fun delete(moduleMatcher: ModuleMatcher): Mono<Int> {
-    return fileDao.findList(moduleMatcher = moduleMatcher).collectList().flatMap { list ->
-      fileDao.delete(moduleMatcher).flatMap {
-        for (fileStore in list) {
-          val file = Paths.get(baseDir, fileStore.path).toFile()
-          if (file.delete()) logger.info("delete file '{}'", file.absolutePath)
-        }
-
-        Mono.just(it)
-      }
-    }
+    return fileAuthorizer
+      // verify permission
+      .verifyHasPermission(moduleMatcher.module, OPERATION_DELETE)
+      .then(Mono.defer {
+        // find them
+        fileDao.findList(moduleMatcher = moduleMatcher)
+          .collectList()
+          // convert empty list to empty mono
+          .flatMap { if (it.isEmpty()) Mono.empty<List<FileStore>>() else Mono.just(it) }
+          .flatMap { list ->
+            // delete db records
+            fileDao.delete(moduleMatcher)
+              .delayUntil {
+                // delete physical files
+                for (fileStore in list) {
+                  val file = Paths.get(baseDir, fileStore.path).toFile()
+                  if (file.delete()) logger.info("delete file '{}'", file.absolutePath)
+                }
+                Mono.empty<Int>()
+              }
+          }
+          .switchIfEmpty(Mono.just(0))
+      })
   }
 }